Adding Chores

+++ internal/storage/migrations/01_init_users.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id INTEGER NOT NULL PRIMARY KEY,
+    id BIGSERIAL NOT NULL PRIMARY KEY,
     authprovider VARCHAR(32) NOT NULL,
     authprovider_id VARCHAR(128) NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
@@ -9,7 +9,7 @@ CREATE TABLE users (
 CREATE INDEX authprovider_users ON users (authprovider, authprovider_id);
 
 CREATE TABLE users_permissions (
-    id INTEGER NOT NULL PRIMARY KEY,
+    id BIGSERIAL NOT NULL PRIMARY KEY,
     user_id INTEGER NOT NULL,
     role VARCHAR(16) NOT NULL,
     UNIQUE(user_id, role)
